+++ tests/compute/subresources/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "pausing.go",
+        "rbac.go",
         "softreboot.go",
         "subresource_api.go",
     ],
